[Backport] Extend fetch_table_stat() to update db cache v2 (#151) (#158)
* Extend fetch_table_stat() to update db cache v2 (#151)

The db cache stores which databases enables diskquota. Active tables
will be recorded only if they are in those databases. Previously,
we created a new UDF update_diskquota_db_list() to add the current db
to the cache. However, the UDF is install in a wrong database. As a
result, after the user upgrade from a previous version to 1.0.3, the
bgworker does not find the UDF and can do nothing.

This patch fixes the issue by removing update_diskquota_db_list() and
using fetch_table_stat() to update db cache. fetch_table_stat() already
exists since version 1.0.0 so that no new UDF is needed.

This PR is a revision of PR #138 , and depends on #130 to fix a race
a condition that occurs after CREATE EXTENSION.

----

Compared to 9684b635, the test has been modified to use pg_sleep()
instead of wait_for_worker_new_epoch() since 1.x doesn't support it.

The extension created by 1.0.3 will still work with new the new so as
long as the UDF update_diskquota_db_list() is not called. (And it won't
be called in the new so.)

Keep the UDF to make the sql the same with 1.0.3. For an easier 2.x
upgrade design.

A FIXME is added to work around the "dbid not found" problem.

Co-authored-by: Xuebin Su (苏学斌) <12034000+xuebinsu@users.noreply.github.com>
Co-authored-by: Chen Mulong <chenmulong@gmail.com>
(cherry picked from commit 9684b6350deb0a7d3849cf4cd202cd97c9b14882)